refactor: 💡 (viva-ms) syncOfficers
To comply with new typed viva client helper
