+++ services/viva/microservice/src/lambdas/syncOfficers.ts
@@ -6,17 +6,58 @@ import config from '../libs/config';
 import log from '../libs/logs';
 
 import vivaAdapter from '../helpers/vivaAdapterRequestClient';
-import officers from '../helpers/officers';
 
-import { CaseAdministrator } from '../types/caseItem';
-import { VivaOfficer, VivaOfficerType } from '../types/vivaMyPages';
-import { CaseItem } from '../types/caseItem';
+import { VivaOfficerType } from '../types/vivaMyPages';
+import type { CaseAdministrator } from '../types/caseItem';
+import type { VivaOfficer } from '../types/vivaMyPages';
+import type { CaseItem } from '../types/caseItem';
 
-type CaseKeys = Pick<CaseItem, 'SK' | 'PK'>;
+interface CaseKeys {
+  PK: string;
+  SK: string;
+}
+
+interface LambdaDetails {
+  dynamodb: {
+    NewImage: Record<string, AWS.DynamoDB.AttributeValue> | undefined;
+  };
+}
+
+interface LambdaRequest {
+  detail: LambdaDetails;
+}
 
 const allowedOfficerTypes: string[] = [VivaOfficerType.Officer];
 
-async function sendUpdateRequest(keys: CaseKeys, newAdministrators: CaseAdministrator[]) {
+export function createCaseAdministrators(
+  vivaOfficer: VivaOfficer | VivaOfficer[]
+): CaseAdministrator[] {
+  let vivaOfficers: VivaOfficer[] = [];
+
+  if (Array.isArray(vivaOfficer)) {
+    vivaOfficers = [...vivaOfficer];
+  } else {
+    // vivaOfficer is an object when viva applicant has only one officer
+    vivaOfficers.push(vivaOfficer);
+  }
+
+  const vivaAdministrators = vivaOfficers.map(officer => {
+    const { name: complexName, title, mail: email, phone, type } = officer;
+    const name = complexName.replace(/^CN=(.+)\/OU.*$/, `$1`);
+
+    return {
+      name,
+      title,
+      email,
+      phone,
+      type,
+    };
+  });
+
+  return vivaAdministrators;
+}
+
+function updateCaseAdministrators(keys: CaseKeys, newAdministrators: CaseAdministrator[]) {
   const TableName = config.cases.tableName;
 
   const dynamoDbParams = {
@@ -31,49 +72,32 @@ async function sendUpdateRequest(keys: CaseKeys, newAdministrators: CaseAdminist
 }
 
 export interface Dependencies {
-  getVivaOfficers: (personalNumber: string) => Promise<{ officer: VivaOfficer | VivaOfficer[] }>;
-  updateCaseAdministrators: (
-    keys: { PK: string; SK: string },
-    administrators: CaseAdministrator[]
-  ) => Promise<void>;
-}
-
-interface LambdaDetails {
-  dynamodb: {
-    NewImage: Record<string, AWS.DynamoDB.AttributeValue> | undefined;
-  };
-}
-
-interface LambdaRequest {
-  detail: LambdaDetails;
+  getVivaOfficers: (personalNumber: number) => Promise<VivaOfficer | VivaOfficer[]>;
+  updateCase: (keys: CaseKeys, administrators: CaseAdministrator[]) => Promise<void>;
 }
 
 export async function syncOfficers(input: LambdaRequest, dependencies: Dependencies) {
-  if (input.detail.dynamodb.NewImage === undefined) {
-    return null;
+  if (!input.detail.dynamodb.NewImage) {
+    return true;
   }
 
-  const { getVivaOfficers, updateCaseAdministrators } = dependencies;
-
   const unMarshalledCaseData = dynamoDb.unmarshall(input.detail.dynamodb.NewImage);
-
   const { PK, SK, details } = unMarshalledCaseData as CaseItem;
   const caseAdministrators = details?.administrators ?? [];
 
   const personalNumber = PK.substring(5);
-  const getVivaOfficersResult = await getVivaOfficers(personalNumber);
+  const vivaOfficers = await dependencies.getVivaOfficers(+personalNumber);
 
-  const { officer } = getVivaOfficersResult;
-  const parsedVivaOfficers = officers.parseVivaOfficers(officer);
-  const vivaOfficers = parsedVivaOfficers.filter(({ type }) =>
+  const administrators = createCaseAdministrators(vivaOfficers);
+  const currentAdministrators = administrators.filter(({ type }) =>
     allowedOfficerTypes.includes(type.toLowerCase())
   );
 
-  if (deepEqual(vivaOfficers, caseAdministrators)) {
-    return false;
+  if (deepEqual(currentAdministrators, caseAdministrators)) {
+    return true;
   }
 
-  await updateCaseAdministrators({ PK, SK }, vivaOfficers);
+  await dependencies.updateCase({ PK, SK }, currentAdministrators);
 
   return true;
 }
@@ -81,6 +105,6 @@ export async function syncOfficers(input: LambdaRequest, dependencies: Dependenc
 export const main = log.wrap(event => {
   return syncOfficers(event, {
     getVivaOfficers: vivaAdapter.officers.get,
-    updateCaseAdministrators: sendUpdateRequest,
+    updateCase: updateCaseAdministrators,
   });
 });
